[IR][ByteType] Introduced a byte type to LLVM IR
This commit introduces `ByteType` to IR. The type extends `IntegerType`
and has a similar functionality. Added:

- Support for byte bitcode writing/reading.
- Support for byte IR assembly.
- Support for byte in C API.

`bitcast` instruction has been modified to accept a byte type as a destination
type. Having it for zext/sext/trunc is not allowed because it can ruin many
optimisations: zext, sext and trunc are not no-ops! So, for truncation
for example we have:

```llvm
%t0 = trunc i32 %v to i8
%t1 = bitcast i8 %t0 to b8
```

Comparison, arithmetic is currently disallowed. Moreover, to reduce the
number of changes there are no byte constants. Instead, a constant expression
bitcast is used, e.g.

```llvm
; storing a 0 byte
store b8 bitcast (i8 0 to b8), b8* %ptr
```

In SelectionDAG, byte type is simply lowered to integer.

Tests to check the assembly round-trip have been added.