[IR][ByteCast] Introduced a bytecast instruction to LLVM IR
This commit adds a `bytecast` instruction, a special cast that operates on a
byte type and reinterprets it as an integer or a pointer. Casts to other types
are not allowed.

The motivation for this instruction is that the byte type may contain either
an integer, and in this case the `bytecast` is the same as a `bitcast`, or a
pointer value, in which case `bytecast` is similar to `ptrtoint`.

Since byte type is lowered as an integer in SelectionDAG, `bytecast` is
lowered as a noop. For GloablISel, there is a placeholder for the
implementation.

Added:
- Bitcode writing/reading support
- Asm parsing
- Code generation for SelDAG and GlobalISel
- Compatibility with C-API

Added tests for `bytecast` instruction, and extended tests for byte type
`bitcast` instruction. Additionally, due to tests failures tests:
  - llvm/test/Transforms/MergeFunc/call-and-invoke-with-ranges.ll
  - llvm/test/Transforms/MergeFunc/inline-asm.ll
were changed.