[Clang][IR][1] Lowering of char/uchar to byte
This is the first commit that introduces a new lowering for char and unsigned
char to a byte type in LLVM IR. Moreover, due to this change strings (`char*`)
are also modelled as a byte type array.

Note: this commit produces a stable IR output for O0 only! Only SOME tests have
been fixed since the nature of the lowering may change in the subsequent
commits.