[OpenMP][IR] Fixes for OpenMP codegen and tests
This commit fixed OpenMP remarks: now these are byte arrays similarly to
a normal string. All OpenMP tests in `llvm` were fixed to adapt to the change.

Note:
=====
Here, we treat `struct.ident_t` as { ... , b8* } instead of { ... , i8* } so
that storing a string pointer is easier. However, Clang needs to be aware of
this too. Fix this in the subsequent commit.