[IR][Mem* fixes][1] Initial lowering for mem* calls
This is the first commit that implements a fix for optimizations of small
memcpy, memset and memcp. Before, the calls to these functions were lowered
to load/store, store, load/store instructions respectively. This was wrong
and as we have seen - lead to miscompilations!

Now, a simple C program:

```c

int my_memcpy(int* a, int* b) {
  memcpy(a, b, sizeof(int));
  return 0;
}
```

is lowered to byte load/store

```llvm
define i32 @my_memcpy(i32* %a, i32* %b) {
entry:
  %0 = bitcast i32* %b to b32*
  %1 = bitcast i32* %a to b32*
  %2 = load b32, b32* %0, align 1
  store b32 %2, b32* %1, align 1
  ret i32 0
}
```

Again, this commit fixes SOME of the LLVM IR tests. The reason for this
partial support is that there are other optimizations for bytes that are
not implemented yet: e.g. bytecast(bitcast(constant)) -> constant. Hence,
where constants are used it makes no sense to adapt tests with long and
incomprehensive sequences of casts.