[Transforms] Fixed store value bitcast in InstCombine
This commit disables the following InstCombine optimization for bytes:

```llvm
store (bitcast %x), %p -> store %x, (bitcast %p)
```

THe problem with thi soptimization was that since a similar one for load and
a bytecast is disabled, the same object in memory was loaded/stored to with
different types. This breaks AA and loop analysis necessary for vectorization.