[Transforms] Made bytecast vectorizable
This commit teaches loop vectorizer to widen bytecasts. Moreover,
the cost of bytecast is modelled as 0, since it is just a reinterpretation
of bytes. Bitcast to byte has also a cost of 0.

This allows to have the same vectorization for `char *`-using loops as for
a version of LLVM without a byte type.