[Transforms][SLP] Added bytecast support
This commit adds bytecast support for SLP vectorizer. Before, a
sequence of bytecasts would not be possible to combine in a
single SIMF bundle - which led to poor vectorization if vectorization
at all.

Now, bytecasts can be also combine together. Currently, we model the
cost of vectorizing a bytecast as 0 - since a byte to integer
conversion is nothing else but a no-op. However, for some cases (e.g.
to pointer conversion) we may want to have something better.

Note: The cost of a bytecast is encoded directly in SLPVectorize. The
main reason for this is that byttecast is not a member of ISD, and
hence cannot be used in TLI.

Moreover, this commit adds a pattern match for bytecasts. It is needed
for load combining, where we look for
  ```
  zext(bytecast(load))
  ```
sequence.