[Analysis][TTI] Fixed cost analysis for inlining
This commit intends to fix inlining issues introduced by byte and bytecast.

Firstly, bytecasts and bitcasts to bytes are ignored during basic block
analysis in order to keep the number of instructions the same as inn the
original LLVM version (or as close as possible).

Secondly, TTI now recognizes bytecasts as zext/sext operands and bitcasts
as trunc users when calculating `CastContextHint`s. This is necessary to
match patterns like

```llvm
  %b = load b8, b8* %ptr, align 1
  %i = bytecast b8 %b to i8
  %v = zext i8 %i to i32
```

where bytecast is a no-op, and the actual operand is a load.