[Clang] Fixed character visitor when src type is a byte
It turns out that AST for C/C++ expressions might be different, for
example:

```c
int foo(unsigned char *c) {
  return c == 'c';
}
```

This program if compiled as a .cpp and with `clang++` has an extra
`ImplicitIntegerCast` node over the character `'c'`. This previously
led to occasional failures on the Clang's side.

This commit adds a type check in `VisitCharacterLiteral` method to
add an extra cast if the actual type of an expression is a byte type.