[IR] Added a byte type guard when comparing with null
This commit fixes a bug where byte comparison was incorrectly created.
While byte comparison is something we want to enable later on, for now
it must be disabled.

The bug originated from Clang's `EmitIntToBoolConversion()` function,
that calls `CreateIsNotNull()`. The latter function simply creates an
`icmp` instruction, no matter what type has been passed to it. To avoid
the issues for Clang's code generation, as well as with other users of
this function and a similar `CreateIsNull()`, the Builder now checks
tha gtype of the passed argument. If it is a byte, an additional bytecast
is created.