[Transforms][LoopIdiom] Added conversion for bytes to memset
This commit fixes an issue where a sequence of byte stores would not
has been recognised as `memset`. The problem was `isBytewiseValue()`,
which did not account for byte values. As a result, only zero bytes
could have been converted to `memset`. Now, any stored byte contant
expression can be converted to `memset`.