[IR] Separated byte and integer strings
Currently, old versions of LLVM and the version with a byte type are not
compatible. The primary reason for this is that a string in my version is
*always* a `b8` array, whereas in old LLVM version it is an array of `i8`.
This makes impossible to parse global variables, defined as:

```llvm
; New LLVM
@var = global constant [6 x b8] c"Hello\00"

; Old LLVM
@var = global constant [6 x i8] c"Hello\00"
```

This commit proposes a solution to this problem by allowing two types of
strings: byte and integer ones. In the subsequent commits, Clang will be changed
to produce a byte type string only for C and C++ frontends.

Now, LLVM IR above is:

```llvm
; New LLVM
@var = global constant [6 x b8] b"Hello\00"

; Old LLVM
@var = global constant [6 x i8] c"Hello\00"
```